requirements.txt
flask==2.3.3
flask-wtf==1.2.1
flask-bootstrap==3.3.7.1
python-dotenv==1.0.0
pinecone-client==3.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.9
langchain-community>=0.0.26
google-generativeai==0.3.1
bs4==0.0.1
requests==2.31.0
pytesseract==0.3.10
Pillow>=8.0.0,<=9.5.0
python-pptx==0.6.22
python-docx==1.0.1
fpdf2==2.7.5
langchain-core>=0.1.29,<0.2.0
langchain-text-splitters==0.0.1
SpeechRecognition==3.10.1
gTTS==2.5.1
pygame==2.5.2
PyAudio==0.2.14
pypdf>=3.0.0